requirements.txt
# Core dependencies
streamlit>=1.28.0
groq>=0.4.0
python-pptx>=0.6.21

# Utilities
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# Optional: For advanced diagram generation
 matplotlib>=3.7.0
plotly>=5.15.0